+++ src/faker_file/contrib/docx_file.py
@@ -1,15 +1,25 @@
 from io import BytesIO
 
+from docx.enum.text import WD_PARAGRAPH_ALIGNMENT
+
 from ..base import DEFAULT_FORMAT_FUNC
 
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
 __copyright__ = "2022-2023 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = (
+    "add_h1_heading",
+    "add_h2_heading",
+    "add_h3_heading",
+    "add_h4_heading",
+    "add_h5_heading",
+    "add_h6_heading",
+    "add_heading",
     "add_page_break",
     "add_paragraph",
     "add_picture",
     "add_table",
+    "add_title_heading",
 )
 
 
@@ -80,3 +90,67 @@ def add_paragraph(provider, document, data, counter, **kwargs):
     data["content_modifiers"]["add_paragraph"].setdefault(counter, [])
     data["content_modifiers"]["add_paragraph"][counter].append(_content)
     data["content"] += "\r\n" + _content
+
+
+def add_heading(provider, document, data, counter, **kwargs):
+    """Callable responsible for the heading generation."""
+    content = kwargs.get("content", None)
+    max_nb_chars = kwargs.get("content", 30)
+    wrap_chars_after = kwargs.get("wrap_chars_after", None)
+    format_func = kwargs.get("format_func", DEFAULT_FORMAT_FUNC)
+    level = kwargs.get("level", 0)
+
+    _content = provider._generate_text_content(
+        max_nb_chars=max_nb_chars,
+        wrap_chars_after=wrap_chars_after,
+        content=content,
+        format_func=format_func,
+    )
+
+    heading = document.add_heading(level=level)
+    heading.add_run(_content)
+
+    # Aligning the heading to center
+    heading.alignment = WD_PARAGRAPH_ALIGNMENT.CENTER
+
+    # Meta-data
+    data.setdefault("content_modifiers", {})
+    data["content_modifiers"].setdefault("add_heading", {})
+    data["content_modifiers"]["add_heading"].setdefault(counter, [])
+    data["content_modifiers"]["add_heading"][counter].append(_content)
+    data["content"] += "\r\n" + _content
+
+
+def add_title_heading(provider, document, data, counter, **kwargs):
+    """Callable responsible for the title heading generation."""
+    return add_heading(provider, document, data, counter, level=0, **kwargs)
+
+
+def add_h1_heading(provider, document, data, counter, **kwargs):
+    """Callable responsible for the h1 heading generation."""
+    return add_heading(provider, document, data, counter, level=1, **kwargs)
+
+
+def add_h2_heading(provider, document, data, counter, **kwargs):
+    """Callable responsible for the h2 heading generation."""
+    return add_heading(provider, document, data, counter, level=2, **kwargs)
+
+
+def add_h3_heading(provider, document, data, counter, **kwargs):
+    """Callable responsible for the h3 heading generation."""
+    return add_heading(provider, document, data, counter, level=3, **kwargs)
+
+
+def add_h4_heading(provider, document, data, counter, **kwargs):
+    """Callable responsible for the h4 heading generation."""
+    return add_heading(provider, document, data, counter, level=4, **kwargs)
+
+
+def add_h5_heading(provider, document, data, counter, **kwargs):
+    """Callable responsible for the h5 heading generation."""
+    return add_heading(provider, document, data, counter, level=5, **kwargs)
+
+
+def add_h6_heading(provider, document, data, counter, **kwargs):
+    """Callable responsible for the h6 heading generation."""
+    return add_heading(provider, document, data, counter, level=6, **kwargs)