+++ src/faker_file/contrib/pdf_file/pdfkit_snippets.py
@@ -6,6 +6,13 @@ __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
 __copyright__ = "2022-2023 Artur Barseghyan"
 __license__ = "MIT"
 __all__ = (
+    "add_h1_heading",
+    "add_h2_heading",
+    "add_h3_heading",
+    "add_h4_heading",
+    "add_h5_heading",
+    "add_h6_heading",
+    "add_heading",
     "add_page_break",
     "add_paragraph",
     "add_picture",
@@ -112,3 +119,80 @@ def add_paragraph(
     data["content_modifiers"]["add_paragraph"].setdefault(counter, [])
     data["content_modifiers"]["add_paragraph"][counter].append(_content)
     data["content"] += "\r\n" + _content
+
+
+def add_heading(
+    provider,
+    generator,
+    document,
+    data,
+    counter,
+    **kwargs,
+):
+    """Callable responsible for heading generation using pdfkit."""
+    content = kwargs.get("content", None)
+    max_nb_chars = kwargs.get("content", 30)
+    wrap_chars_after = kwargs.get("wrap_chars_after", None)
+    format_func = kwargs.get("format_func", DEFAULT_FORMAT_FUNC)
+    level = kwargs.get("level", 1)
+    if level < 1 or level > 6:
+        level = 1
+
+    _content = provider._generate_text_content(
+        max_nb_chars=max_nb_chars,
+        wrap_chars_after=wrap_chars_after,
+        content=content,
+        format_func=format_func,
+    )
+
+    heading_html = f"<div><h{level}>{_content}</h{level}></div>"
+    document += "\r\n" + heading_html
+
+    # Meta-data
+    data.setdefault("content_modifiers", {})
+    data["content_modifiers"].setdefault("add_heading", {})
+    data["content_modifiers"]["add_heading"].setdefault(counter, [])
+    data["content_modifiers"]["add_heading"][counter].append(_content)
+    data["content"] += "\r\n" + _content
+
+
+def add_h1_heading(provider, generator, document, data, counter, **kwargs):
+    """Callable responsible for the h1 heading generation."""
+    return add_heading(
+        provider, generator, document, data, counter, level=1, **kwargs
+    )
+
+
+def add_h2_heading(provider, generator, document, data, counter, **kwargs):
+    """Callable responsible for the h2 heading generation."""
+    return add_heading(
+        provider, generator, document, data, counter, level=2, **kwargs
+    )
+
+
+def add_h3_heading(provider, generator, document, data, counter, **kwargs):
+    """Callable responsible for the h3 heading generation."""
+    return add_heading(
+        provider, generator, document, data, counter, level=3, **kwargs
+    )
+
+
+def add_h4_heading(provider, generator, document, data, counter, **kwargs):
+    """Callable responsible for the h4 heading generation."""
+    return add_heading(
+        provider, generator, document, data, counter, level=4, **kwargs
+    )
+
+
+def add_h5_heading(provider, generator, document, data, counter, **kwargs):
+    """Callable responsible for the h5 heading generation."""
+    return add_heading(
+        provider, generator, document, data, counter, level=5, **kwargs
+    )
+
+
+def add_h6_heading(provider, generator, document, data, counter, **kwargs):
+    """Callable responsible for the h6 heading generation."""
+    return add_heading(
+        provider, generator, document, data, counter, level=6, **kwargs
+    )